refactor: 고객 banking section 컴포넌트 분리
