+++ front/src/app/page.tsx
@@ -1,6 +1,12 @@
 "use client";
 
 import type { FormEvent } from "react";
+import { AccountsSection } from "@/components/customer-banking/sections/accounts-section";
+import { DashboardSection } from "@/components/customer-banking/sections/dashboard-section";
+import { NotificationsSection } from "@/components/customer-banking/sections/notifications-section";
+import { SecuritySection } from "@/components/customer-banking/sections/security-section";
+import { TransactionsSection } from "@/components/customer-banking/sections/transactions-section";
+import { TransferSection } from "@/components/customer-banking/sections/transfer-section";
 import { useEffect, useMemo, useRef, useState } from "react";
 import { AquilaBankApiClient } from "@/lib/api/client";
 import {
@@ -1002,1687 +1008,5 @@ export default function HomePage() {
         </aside>
       </div>
     </main>
-  );
-}
-
-function DashboardSection({
-  hasSession,
-  isBusy,
-  onMove,
-  onRefresh,
-}: {
-  hasSession: boolean;
-  isBusy: boolean;
-  onMove: (section: MenuSection) => void;
-  onRefresh: () => void;
-}) {
-  return (
-    <section className="task-section">
-      <div className="section-title">
-        <div>
-          <p>개인뱅킹</p>
-          <h1>뱅킹 업무</h1>
-        </div>
-        <div className="button-row">
-          <button onClick={() => onMove("accounts")} type="button">
-            계좌조회
-          </button>
-          <button onClick={() => onMove("transfer")} type="button">
-            이체
-          </button>
-          <button disabled={!hasSession || isBusy} onClick={onRefresh} type="button">
-            토큰 재발급
-          </button>
-        </div>
-      </div>
-      <div className="summary-grid">
-        <article className="summary-box">
-          <span>조회</span>
-          <strong>계좌 목록/잔액</strong>
-          <button onClick={() => onMove("accounts")} type="button">
-            바로가기
-          </button>
-        </article>
-        <article className="summary-box">
-          <span>이체</span>
-          <strong>즉시이체/취소</strong>
-          <button onClick={() => onMove("transfer")} type="button">
-            바로가기
-          </button>
-        </article>
-        <article className="summary-box">
-          <span>조회</span>
-          <strong>거래내역</strong>
-          <button onClick={() => onMove("transactions")} type="button">
-            바로가기
-          </button>
-        </article>
-        <article className="summary-box">
-          <span>알림</span>
-          <strong>알림함/설정</strong>
-          <button onClick={() => onMove("notifications")} type="button">
-            바로가기
-          </button>
-        </article>
-      </div>
-      <div className="bank-table-wrap">
-        <table className="bank-table">
-          <caption>업무별 처리 기준</caption>
-          <thead>
-            <tr>
-              <th>업무</th>
-              <th>기준</th>
-              <th>상태</th>
-            </tr>
-          </thead>
-          <tbody>
-            <tr>
-              <td>거래내역 조회</td>
-              <td>계좌, 기간, limit, cursor 기반</td>
-              <td>bounded query</td>
-            </tr>
-            <tr>
-              <td>이체</td>
-              <td>요청 단위 Idempotency-Key 발급</td>
-              <td>중복 방지</td>
-            </tr>
-            <tr>
-              <td>알림</td>
-              <td>SSE 연결 상태와 inbox 동시 제공</td>
-              <td>실시간</td>
-            </tr>
-          </tbody>
-        </table>
-      </div>
-    </section>
-  );
-}
-
-function AccountsSection({
-  accountCursor,
-  accountLimit,
-  accounts,
-  isBusy,
-  selectedAccount,
-  onAccountLimitChange,
-  onLoadAccountDetail,
-  onLoadAccounts,
-  onLoadNextAccounts,
-}: {
-  accountCursor: string;
-  accountLimit: number;
-  accounts: AccountItem[];
-  isBusy: boolean;
-  selectedAccount: AccountSummaryResponse | null;
-  onAccountLimitChange: (value: number) => void;
-  onLoadAccountDetail: (accountId: number) => void;
-  onLoadAccounts: () => void;
-  onLoadNextAccounts: () => void;
-}) {
-  return (
-    <section className="task-section">
-      <div className="section-title">
-        <div>
-          <p>조회</p>
-          <h1>계좌조회</h1>
-        </div>
-        <div className="button-row">
-          <label className="inline-control">
-            <span>건수</span>
-            <select
-              onChange={(event) => onAccountLimitChange(Number(event.target.value))}
-              value={accountLimit}
-            >
-              <option value={10}>10</option>
-              <option value={20}>20</option>
-              <option value={50}>50</option>
-            </select>
-          </label>
-          <button disabled={isBusy} onClick={onLoadAccounts} type="button">
-            조회
-          </button>
-          <button disabled={!accountCursor || isBusy} onClick={onLoadNextAccounts} type="button">
-            다음
-          </button>
-        </div>
-      </div>
-
-      <div className="account-grid">
-        <section className="table-panel embedded">
-          <div className="panel-toolbar">
-            <div>
-              <strong>보유계좌</strong>
-              <span>계좌별 잔액과 상태</span>
-            </div>
-          </div>
-          <div className="bank-table-wrap">
-            <table className="bank-table">
-              <caption>계좌 목록</caption>
-              <thead>
-                <tr>
-                  <th>계좌번호</th>
-                  <th>계좌명</th>
-                  <th>상태</th>
-                  <th>출금가능액</th>
-                  <th>관리</th>
-                </tr>
-              </thead>
-              <tbody>
-                {accounts.length === 0 ? (
-                  <tr>
-                    <td colSpan={5}>조회된 계좌가 없습니다.</td>
-                  </tr>
-                ) : (
-                  accounts.map((account) => (
-                    <tr key={account.accountId}>
-                      <td>{account.accountNumber}</td>
-                      <td>{account.displayName}</td>
-                      <td>
-                        <span className="status-badge">{account.accountStatus}</span>
-                      </td>
-                      <td className="amount-cell">
-                        {formatMinorAmount(
-                          account.availableBalanceMinor,
-                          account.currencyCode,
-                        )}
-                      </td>
-                      <td>
-                        <button
-                          disabled={isBusy}
-                          onClick={() => onLoadAccountDetail(account.accountId)}
-                          type="button"
-                        >
-                          상세
-                        </button>
-                      </td>
-                    </tr>
-                  ))
-                )}
-              </tbody>
-            </table>
-          </div>
-        </section>
-
-        <aside className="detail-panel">
-          <div className="form-heading">
-            <strong>계좌상세</strong>
-            <span>{selectedAccount ? selectedAccount.accountStatus : "미선택"}</span>
-          </div>
-          {selectedAccount ? (
-            <>
-              <div className="balance-card">
-                <span>{selectedAccount.displayName}</span>
-                <strong>
-                  {formatMinorAmount(
-                    selectedAccount.availableBalanceMinor,
-                    selectedAccount.currencyCode,
-                  )}
-                </strong>
-                <small>출금가능금액</small>
-              </div>
-              <dl className="detail-list">
-                <div>
-                  <dt>Account ID</dt>
-                  <dd>{selectedAccount.accountId}</dd>
-                </div>
-                <div>
-                  <dt>계좌번호</dt>
-                  <dd>{selectedAccount.accountNumber}</dd>
-                </div>
-                <div>
-                  <dt>보류금액</dt>
-                  <dd>
-                    {formatMinorAmount(
-                      selectedAccount.pendingBalanceMinor,
-                      selectedAccount.currencyCode,
-                    )}
-                  </dd>
-                </div>
-                <div>
-                  <dt>잔액 갱신</dt>
-                  <dd>{formatDateTime(selectedAccount.balanceUpdatedAt)}</dd>
-                </div>
-              </dl>
-            </>
-          ) : (
-            <p className="rail-copy">계좌 목록에서 상세 버튼을 선택하세요.</p>
-          )}
-        </aside>
-      </div>
-    </section>
-  );
-}
-
-function TransferSection({
-  isBusy,
-  reversalForm,
-  reversalResult,
-  transferForm,
-  transferResult,
-  onReversal,
-  onReversalChange,
-  onTransfer,
-  onTransferChange,
-}: {
-  isBusy: boolean;
-  reversalForm: {
-    transactionReference: string;
-    sourceAccountId: string;
-    amountMinor: string;
-    reversalReason: string;
-    summary: string;
-  };
-  reversalResult: TransferReversalResponse | null;
-  transferForm: {
-    sourceAccountId: string;
-    targetAccountId: string;
-    amountMinor: string;
-    currencyCode: string;
-    summary: string;
-  };
-  transferResult: TransferResponse | null;
-  onReversal: (event: FormEvent<HTMLFormElement>) => void;
-  onReversalChange: (value: {
-    transactionReference: string;
-    sourceAccountId: string;
-    amountMinor: string;
-    reversalReason: string;
-    summary: string;
-  }) => void;
-  onTransfer: (event: FormEvent<HTMLFormElement>) => void;
-  onTransferChange: (value: {
-    sourceAccountId: string;
-    targetAccountId: string;
-    amountMinor: string;
-    currencyCode: string;
-    summary: string;
-  }) => void;
-}) {
-  return (
-    <section className="task-section">
-      <div className="section-title">
-        <div>
-          <p>이체</p>
-          <h1>즉시이체</h1>
-        </div>
-      </div>
-
-      <div className="two-column">
-        <form className="bank-form" onSubmit={onTransfer}>
-          <div className="form-heading">
-            <strong>이체정보 입력</strong>
-            <span>요청 단위 중복 방지</span>
-          </div>
-          <div className="form-grid">
-            <label>
-              <span>출금계좌 ID</span>
-              <input
-                inputMode="numeric"
-                onChange={(event) =>
-                  onTransferChange({
-                    ...transferForm,
-                    sourceAccountId: event.target.value,
-                  })
-                }
-                required
-                value={transferForm.sourceAccountId}
-              />
-            </label>
-            <label>
-              <span>입금계좌 ID</span>
-              <input
-                inputMode="numeric"
-                onChange={(event) =>
-                  onTransferChange({
-                    ...transferForm,
-                    targetAccountId: event.target.value,
-                  })
-                }
-                required
-                value={transferForm.targetAccountId}
-              />
-            </label>
-            <label>
-              <span>금액 minor</span>
-              <input
-                inputMode="numeric"
-                onChange={(event) =>
-                  onTransferChange({
-                    ...transferForm,
-                    amountMinor: event.target.value,
-                  })
-                }
-                required
-                value={transferForm.amountMinor}
-              />
-            </label>
-            <label>
-              <span>통화</span>
-              <input
-                maxLength={3}
-                onChange={(event) =>
-                  onTransferChange({
-                    ...transferForm,
-                    currencyCode: event.target.value.toUpperCase(),
-                  })
-                }
-                required
-                value={transferForm.currencyCode}
-              />
-            </label>
-          </div>
-          <label>
-            <span>받는 분 통장 표시</span>
-            <input
-              maxLength={120}
-              onChange={(event) =>
-                onTransferChange({ ...transferForm, summary: event.target.value })
-              }
-              required
-              value={transferForm.summary}
-            />
-          </label>
-          <button disabled={isBusy} type="submit">
-            이체 실행
-          </button>
-        </form>
-
-        <ResultPanel
-          title="이체 결과"
-          rows={
-            transferResult
-              ? [
-                  ["거래번호", transferResult.transactionReference],
-                  ["상태", transferResult.status],
-                  [
-                    "이체금액",
-                    formatMinorAmount(
-                      transferResult.amountMinor,
-                      transferResult.currencyCode,
-                    ),
-                  ],
-                  [
-                    "이체 후 잔액",
-                    formatMinorAmount(
-                      transferResult.availableBalanceAfterMinor,
-                      transferResult.currencyCode,
-                    ),
-                  ],
-                  ["기장시각", formatDateTime(transferResult.bookedAt)],
-                ]
-              : []
-          }
-        />
-      </div>
-
-      <div className="two-column">
-        <form className="bank-form" onSubmit={onReversal}>
-          <div className="form-heading">
-            <strong>이체 취소</strong>
-            <span>원거래 reference 기준</span>
-          </div>
-          <label>
-            <span>원거래번호</span>
-            <input
-              onChange={(event) =>
-                onReversalChange({
-                  ...reversalForm,
-                  transactionReference: event.target.value,
-                })
-              }
-              required
-              value={reversalForm.transactionReference}
-            />
-          </label>
-          <div className="form-grid">
-            <label>
-              <span>출금계좌 ID</span>
-              <input
-                inputMode="numeric"
-                onChange={(event) =>
-                  onReversalChange({
-                    ...reversalForm,
-                    sourceAccountId: event.target.value,
-                  })
-                }
-                required
-                value={reversalForm.sourceAccountId}
-              />
-            </label>
-            <label>
-              <span>취소금액 minor</span>
-              <input
-                inputMode="numeric"
-                onChange={(event) =>
-                  onReversalChange({
-                    ...reversalForm,
-                    amountMinor: event.target.value,
-                  })
-                }
-                required
-                value={reversalForm.amountMinor}
-              />
-            </label>
-          </div>
-          <label>
-            <span>취소사유</span>
-            <select
-              onChange={(event) =>
-                onReversalChange({
-                  ...reversalForm,
-                  reversalReason: event.target.value,
-                })
-              }
-              value={reversalForm.reversalReason}
-            >
-              <option value="CUSTOMER_REQUEST">CUSTOMER_REQUEST</option>
-              <option value="DUPLICATE">DUPLICATE</option>
-              <option value="WRONG_AMOUNT">WRONG_AMOUNT</option>
-              <option value="WRONG_TARGET">WRONG_TARGET</option>
-              <option value="FRAUD_REPORTED">FRAUD_REPORTED</option>
-            </select>
-          </label>
-          <label>
-            <span>적요</span>
-            <input
-              maxLength={120}
-              onChange={(event) =>
-                onReversalChange({ ...reversalForm, summary: event.target.value })
-              }
-              required
-              value={reversalForm.summary}
-            />
-          </label>
-          <button disabled={isBusy} type="submit">
-            취소 실행
-          </button>
-        </form>
-
-        <ResultPanel
-          title="취소 결과"
-          rows={
-            reversalResult
-              ? [
-                  ["원거래", reversalResult.originalTransactionReference],
-                  ["취소거래", reversalResult.reversalTransactionReference],
-                  ["상태", reversalResult.status],
-                  [
-                    "취소금액",
-                    formatMinorAmount(
-                      reversalResult.amountMinor,
-                      reversalResult.currencyCode,
-                    ),
-                  ],
-                  ["기장시각", formatDateTime(reversalResult.bookedAt)],
-                ]
-              : []
-          }
-        />
-      </div>
-    </section>
-  );
-}
-
-function TransactionsSection({
-  filters,
-  isBusy,
-  mode,
-  slice,
-  transactionDetail,
-  transactions,
-  onDetail,
-  onFilterChange,
-  onModeChange,
-  onNext,
-  onSearch,
-}: {
-  filters: {
-    accountId: string;
-    from: string;
-    to: string;
-    limit: string;
-    cursor: string;
-    status: string;
-    direction: string;
-    minAmountMinor: string;
-    maxAmountMinor: string;
-    transactionReference: string;
-    responseShape: "full" | "slim";
-  };
-  isBusy: boolean;
-  mode: "active" | "archive";
-  slice: TransactionQueryResponse | null;
-  transactionDetail: TransactionDetailResponse | null;
-  transactions: TransactionItem[];
-  onDetail: (transactionReference: string) => void;
-  onFilterChange: (value: {
-    accountId: string;
-    from: string;
-    to: string;
-    limit: string;
-    cursor: string;
-    status: string;
-    direction: string;
-    minAmountMinor: string;
-    maxAmountMinor: string;
-    transactionReference: string;
-    responseShape: "full" | "slim";
-  }) => void;
-  onModeChange: (mode: "active" | "archive") => void;
-  onNext: () => void;
-  onSearch: (event: FormEvent<HTMLFormElement>) => void;
-}) {
-  return (
-    <section className="task-section">
-      <div className="section-title">
-        <div>
-          <p>조회</p>
-          <h1>거래내역 조회</h1>
-        </div>
-        <div className="tab-switch" role="tablist" aria-label="거래 조회 구분">
-          <button
-            aria-selected={mode === "active"}
-            className={mode === "active" ? "active" : ""}
-            onClick={() => onModeChange("active")}
-            role="tab"
-            type="button"
-          >
-            일반
-          </button>
-          <button
-            aria-selected={mode === "archive"}
-            className={mode === "archive" ? "active" : ""}
-            onClick={() => onModeChange("archive")}
-            role="tab"
-            type="button"
-          >
-            아카이브
-          </button>
-        </div>
-      </div>
-
-      <form className="bank-form filter-form" onSubmit={onSearch}>
-        <div className="filter-grid">
-          <label>
-            <span>계좌 ID</span>
-            <input
-              inputMode="numeric"
-              onChange={(event) =>
-                onFilterChange({ ...filters, accountId: event.target.value })
-              }
-              required
-              value={filters.accountId}
-            />
-          </label>
-          <label>
-            <span>시작일시</span>
-            <input
-              onChange={(event) =>
-                onFilterChange({ ...filters, from: event.target.value })
-              }
-              required
-              type="datetime-local"
-              value={filters.from}
-            />
-          </label>
-          <label>
-            <span>종료일시</span>
-            <input
-              onChange={(event) =>
-                onFilterChange({ ...filters, to: event.target.value })
-              }
-              required
-              type="datetime-local"
-              value={filters.to}
-            />
-          </label>
-          <label>
-            <span>건수</span>
-            <select
-              onChange={(event) =>
-                onFilterChange({ ...filters, limit: event.target.value })
-              }
-              value={filters.limit}
-            >
-              <option value="20">20</option>
-              <option value="50">50</option>
-              <option value="100">100</option>
-            </select>
-          </label>
-          <label>
-            <span>상태</span>
-            <select
-              onChange={(event) =>
-                onFilterChange({ ...filters, status: event.target.value })
-              }
-              value={filters.status}
-            >
-              <option value="">전체</option>
-              <option value="PENDING">PENDING</option>
-              <option value="BOOKED">BOOKED</option>
-              <option value="REVERSED">REVERSED</option>
-              <option value="FAILED">FAILED</option>
-            </select>
-          </label>
-          <label>
-            <span>입출금</span>
-            <select
-              onChange={(event) =>
-                onFilterChange({ ...filters, direction: event.target.value })
-              }
-              value={filters.direction}
-            >
-              <option value="">전체</option>
-              <option value="DEBIT">출금</option>
-              <option value="CREDIT">입금</option>
-            </select>
-          </label>
-          <label>
-            <span>최소금액 minor</span>
-            <input
-              inputMode="numeric"
-              onChange={(event) =>
-                onFilterChange({
-                  ...filters,
-                  minAmountMinor: event.target.value,
-                })
-              }
-              value={filters.minAmountMinor}
-            />
-          </label>
-          <label>
-            <span>최대금액 minor</span>
-            <input
-              inputMode="numeric"
-              onChange={(event) =>
-                onFilterChange({
-                  ...filters,
-                  maxAmountMinor: event.target.value,
-                })
-              }
-              value={filters.maxAmountMinor}
-            />
-          </label>
-          <label>
-            <span>거래번호</span>
-            <input
-              onChange={(event) =>
-                onFilterChange({
-                  ...filters,
-                  transactionReference: event.target.value,
-                })
-              }
-              value={filters.transactionReference}
-            />
-          </label>
-          <label>
-            <span>응답형태</span>
-            <select
-              onChange={(event) =>
-                onFilterChange({
-                  ...filters,
-                  responseShape: event.target.value as "full" | "slim",
-                })
-              }
-              value={filters.responseShape}
-            >
-              <option value="full">full</option>
-              <option value="slim">slim</option>
-            </select>
-          </label>
-        </div>
-        <div className="button-row">
-          <button disabled={isBusy} type="submit">
-            조회
-          </button>
-          <button disabled={!slice?.nextCursor || isBusy} onClick={onNext} type="button">
-            다음 거래
-          </button>
-        </div>
-      </form>
-
-      <div className="split-work">
-        <section className="table-panel embedded">
-          <div className="panel-toolbar">
-            <div>
-              <strong>거래내역</strong>
-              <span>
-                {slice
-                  ? `${transactions.length}건 표시 / next ${slice.hasNext ? "있음" : "없음"}`
-                  : "조회 전"}
-              </span>
-            </div>
-          </div>
-          <div className="bank-table-wrap">
-            <table className="bank-table">
-              <caption>거래내역 목록</caption>
-              <thead>
-                <tr>
-                  <th>기장일시</th>
-                  <th>거래번호</th>
-                  <th>입출금</th>
-                  <th>상태</th>
-                  <th>금액</th>
-                  <th>잔액</th>
-                  <th>상세</th>
-                </tr>
-              </thead>
-              <tbody>
-                {transactions.length === 0 ? (
-                  <tr>
-                    <td colSpan={7}>조회된 거래가 없습니다.</td>
-                  </tr>
-                ) : (
-                  transactions.map((item) => (
-                    <tr key={`${item.id}-${item.transactionReference}`}>
-                      <td>{formatDateTime(item.bookedAt)}</td>
-                      <td>{item.transactionReference}</td>
-                      <td>{item.direction}</td>
-                      <td>
-                        <span className="status-badge">{item.status}</span>
-                      </td>
-                      <td className={item.direction === "DEBIT" ? "amount debit" : "amount credit"}>
-                        {formatMinorAmount(item.amountMinor, item.currencyCode)}
-                      </td>
-                      <td>
-                        {item.balanceAfterMinor == null
-                          ? "-"
-                          : formatMinorAmount(
-                              item.balanceAfterMinor,
-                              item.currencyCode,
-                            )}
-                      </td>
-                      <td>
-                        <button
-                          disabled={isBusy}
-                          onClick={() => onDetail(item.transactionReference)}
-                          type="button"
-                        >
-                          상세
-                        </button>
-                      </td>
-                    </tr>
-                  ))
-                )}
-              </tbody>
-            </table>
-          </div>
-        </section>
-
-        <aside className="detail-panel">
-          <div className="form-heading">
-            <strong>거래상세</strong>
-            <span>{transactionDetail?.transactionStatus ?? "미선택"}</span>
-          </div>
-          {transactionDetail ? (
-            <dl className="detail-list">
-              <div>
-                <dt>거래번호</dt>
-                <dd>{transactionDetail.transactionReference}</dd>
-              </div>
-              <div>
-                <dt>입출금</dt>
-                <dd>{transactionDetail.direction}</dd>
-              </div>
-              <div>
-                <dt>금액</dt>
-                <dd>
-                  {formatMinorAmount(
-                    transactionDetail.amountMinor,
-                    transactionDetail.currencyCode,
-                  )}
-                </dd>
-              </div>
-              <div>
-                <dt>잔액</dt>
-                <dd>
-                  {formatMinorAmount(
-                    transactionDetail.balanceAfterMinor,
-                    transactionDetail.currencyCode,
-                  )}
-                </dd>
-              </div>
-              <div>
-                <dt>상대방</dt>
-                <dd>{transactionDetail.counterpartyMaskedName ?? "-"}</dd>
-              </div>
-              <div>
-                <dt>Ledger</dt>
-                <dd>{transactionDetail.entryReference}</dd>
-              </div>
-              <div>
-                <dt>발생시각</dt>
-                <dd>{formatDateTime(transactionDetail.occurredAt)}</dd>
-              </div>
-              <div>
-                <dt>설명</dt>
-                <dd>{transactionDetail.description ?? "-"}</dd>
-              </div>
-            </dl>
-          ) : (
-            <p className="rail-copy">거래내역에서 상세 버튼을 선택하세요.</p>
-          )}
-        </aside>
-      </div>
-    </section>
-  );
-}
-
-function ResultPanel({ title, rows }: { title: string; rows: string[][] }) {
-  return (
-    <div className="result-panel">
-      <div className="form-heading">
-        <strong>{title}</strong>
-        <span>{rows.length > 0 ? "완료" : "대기"}</span>
-      </div>
-      {rows.length === 0 ? (
-        <p className="rail-copy">처리 결과가 여기에 표시됩니다.</p>
-      ) : (
-        <dl className="detail-list">
-          {rows.map(([key, value]) => (
-            <div key={key}>
-              <dt>{key}</dt>
-              <dd>{value}</dd>
-            </div>
-          ))}
-        </dl>
-      )}
-    </div>
-  );
-}
-
-function NotificationsSection({
-  filters,
-  isBusy,
-  mode,
-  notifications,
-  preferences,
-  selectedIds,
-  slice,
-  sseStatus,
-  unreadCount,
-  onBulkAction,
-  onConnect,
-  onDisconnect,
-  onFilterChange,
-  onLoadPreferences,
-  onLoadUnreadCount,
-  onModeChange,
-  onNext,
-  onPreferenceChange,
-  onSearch,
-  onToggleId,
-  onUpdatePreferences,
-}: {
-  filters: {
-    limit: string;
-    cursor: string;
-    readStatus: string;
-    eventType: string;
-    from: string;
-    to: string;
-  };
-  isBusy: boolean;
-  mode: "inbox" | "search";
-  notifications: NotificationItem[];
-  preferences: NotificationPreferenceItem[];
-  selectedIds: number[];
-  slice: NotificationQueryResponse | null;
-  sseStatus: {
-    state: string;
-    lastEventAt: string;
-    lastEventId: string;
-  };
-  unreadCount: number | null;
-  onBulkAction: (
-    action: "read" | "archive" | "delete",
-    notificationId?: number,
-  ) => void;
-  onConnect: () => void;
-  onDisconnect: () => void;
-  onFilterChange: (value: {
-    limit: string;
-    cursor: string;
-    readStatus: string;
-    eventType: string;
-    from: string;
-    to: string;
-  }) => void;
-  onLoadPreferences: () => void;
-  onLoadUnreadCount: () => void;
-  onModeChange: (mode: "inbox" | "search") => void;
-  onNext: () => void;
-  onPreferenceChange: (value: NotificationPreferenceItem[]) => void;
-  onSearch: (event: FormEvent<HTMLFormElement>) => void;
-  onToggleId: (notificationId: number) => void;
-  onUpdatePreferences: () => void;
-}) {
-  return (
-    <section className="task-section">
-      <div className="section-title">
-        <div>
-          <p>고객센터</p>
-          <h1>알림</h1>
-        </div>
-        <div className="button-row">
-          <button disabled={isBusy} onClick={onLoadUnreadCount} type="button">
-            미확인 조회
-          </button>
-          <button disabled={isBusy} onClick={onConnect} type="button">
-            SSE 연결
-          </button>
-          <button onClick={onDisconnect} type="button">
-            연결 해제
-          </button>
-        </div>
-      </div>
-
-      <div className="notification-status">
-        <div>
-          <span>SSE</span>
-          <strong>{sseStatus.state}</strong>
-        </div>
-        <div>
-          <span>Unread</span>
-          <strong>{unreadCount ?? "-"}</strong>
-        </div>
-        <div>
-          <span>Last Event</span>
-          <strong>{sseStatus.lastEventId || "-"}</strong>
-        </div>
-        <div>
-          <span>수신시각</span>
-          <strong>{formatDateTime(sseStatus.lastEventAt)}</strong>
-        </div>
-      </div>
-
-      <form className="bank-form filter-form" onSubmit={onSearch}>
-        <div className="panel-toolbar inline-toolbar">
-          <div className="tab-switch" role="tablist" aria-label="알림 조회 구분">
-            <button
-              aria-selected={mode === "inbox"}
-              className={mode === "inbox" ? "active" : ""}
-              onClick={() => onModeChange("inbox")}
-              role="tab"
-              type="button"
-            >
-              Inbox
-            </button>
-            <button
-              aria-selected={mode === "search"}
-              className={mode === "search" ? "active" : ""}
-              onClick={() => onModeChange("search")}
-              role="tab"
-              type="button"
-            >
-              Search
-            </button>
-          </div>
-          <div className="button-row compact">
-            <button disabled={isBusy} type="submit">
-              조회
-            </button>
-            <button disabled={!slice?.nextCursor || isBusy} onClick={onNext} type="button">
-              다음
-            </button>
-          </div>
-        </div>
-        <div className="filter-grid notification-filter">
-          <label>
-            <span>건수</span>
-            <select
-              onChange={(event) =>
-                onFilterChange({ ...filters, limit: event.target.value })
-              }
-              value={filters.limit}
-            >
-              <option value="20">20</option>
-              <option value="50">50</option>
-              <option value="100">100</option>
-            </select>
-          </label>
-          <label>
-            <span>읽음상태</span>
-            <select
-              disabled={mode === "inbox"}
-              onChange={(event) =>
-                onFilterChange({ ...filters, readStatus: event.target.value })
-              }
-              value={filters.readStatus}
-            >
-              <option value="ALL">ALL</option>
-              <option value="READ">READ</option>
-              <option value="UNREAD">UNREAD</option>
-            </select>
-          </label>
-          <label>
-            <span>이벤트 유형</span>
-            <input
-              disabled={mode === "inbox"}
-              onChange={(event) =>
-                onFilterChange({ ...filters, eventType: event.target.value })
-              }
-              value={filters.eventType}
-            />
-          </label>
-          <label>
-            <span>시작일시</span>
-            <input
-              disabled={mode === "inbox"}
-              onChange={(event) =>
-                onFilterChange({ ...filters, from: event.target.value })
-              }
-              type="datetime-local"
-              value={filters.from}
-            />
-          </label>
-          <label>
-            <span>종료일시</span>
-            <input
-              disabled={mode === "inbox"}
-              onChange={(event) =>
-                onFilterChange({ ...filters, to: event.target.value })
-              }
-              type="datetime-local"
-              value={filters.to}
-            />
-          </label>
-        </div>
-      </form>
-
-      <div className="split-work">
-        <section className="table-panel embedded">
-          <div className="panel-toolbar">
-            <div>
-              <strong>알림함</strong>
-              <span>
-                {slice
-                  ? `${notifications.length}건 표시 / next ${slice.hasNext ? "있음" : "없음"}`
-                  : "조회 전"}
-              </span>
-            </div>
-            <div className="button-row compact">
-              <button
-                disabled={selectedIds.length === 0 || isBusy}
-                onClick={() => onBulkAction("read")}
-                type="button"
-              >
-                읽음
-              </button>
-              <button
-                disabled={selectedIds.length === 0 || isBusy}
-                onClick={() => onBulkAction("archive")}
-                type="button"
-              >
-                보관
-              </button>
-              <button
-                disabled={selectedIds.length === 0 || isBusy}
-                onClick={() => onBulkAction("delete")}
-                type="button"
-              >
-                삭제
-              </button>
-            </div>
-          </div>
-          <div className="bank-table-wrap">
-            <table className="bank-table">
-              <caption>알림 목록</caption>
-              <thead>
-                <tr>
-                  <th>선택</th>
-                  <th>상태</th>
-                  <th>유형</th>
-                  <th>제목</th>
-                  <th>수신시각</th>
-                  <th>관리</th>
-                </tr>
-              </thead>
-              <tbody>
-                {notifications.length === 0 ? (
-                  <tr>
-                    <td colSpan={6}>조회된 알림이 없습니다.</td>
-                  </tr>
-                ) : (
-                  notifications.map((item) => (
-                    <tr key={item.notificationId}>
-                      <td>
-                        <input
-                          aria-label={`${item.notificationId} 선택`}
-                          checked={selectedIds.includes(item.notificationId)}
-                          onChange={() => onToggleId(item.notificationId)}
-                          type="checkbox"
-                        />
-                      </td>
-                      <td>
-                        <span className={item.read ? "status-badge" : "status-badge unread"}>
-                          {item.read ? "READ" : "UNREAD"}
-                        </span>
-                      </td>
-                      <td>{item.eventType}</td>
-                      <td>
-                        <strong>{item.title}</strong>
-                        <p className="table-message">{item.message}</p>
-                      </td>
-                      <td>{formatDateTime(item.createdAt)}</td>
-                      <td>
-                        <button
-                          disabled={item.read || isBusy}
-                          onClick={() => onBulkAction("read", item.notificationId)}
-                          type="button"
-                        >
-                          읽음
-                        </button>
-                      </td>
-                    </tr>
-                  ))
-                )}
-              </tbody>
-            </table>
-          </div>
-        </section>
-
-        <aside className="detail-panel">
-          <div className="form-heading">
-            <strong>알림 수신 설정</strong>
-            <span>{preferences.length}건</span>
-          </div>
-          <div className="button-row compact preference-actions">
-            <button disabled={isBusy} onClick={onLoadPreferences} type="button">
-              설정 조회
-            </button>
-            <button disabled={preferences.length === 0 || isBusy} onClick={onUpdatePreferences} type="button">
-              저장
-            </button>
-          </div>
-          {preferences.length === 0 ? (
-            <p className="rail-copy">알림 설정을 조회하세요.</p>
-          ) : (
-            <div className="preference-list">
-              {preferences.map((item, index) => (
-                <label className="preference-row" key={`${item.category}-${item.channel}`}>
-                  <input
-                    checked={item.enabled}
-                    onChange={(event) => {
-                      const nextItems = [...preferences];
-                      nextItems[index] = {
-                        ...item,
-                        enabled: event.target.checked,
-                      };
-                      onPreferenceChange(nextItems);
-                    }}
-                    type="checkbox"
-                  />
-                  <span>
-                    <strong>{item.category}</strong>
-                    <small>{item.channel}</small>
-                  </span>
-                </label>
-              ))}
-            </div>
-          )}
-        </aside>
-      </div>
-    </section>
-  );
-}
-
-function SecuritySection(props: {
-  backupChallengeForm: {
-    backupCode: string;
-    rememberDevice: boolean;
-  };
-  backupCodes: BackupCodeIssueResponse | null;
-  challenge: LoginResponse | null;
-  isBusy: boolean;
-  loginForm: {
-    loginId: string;
-    password: string;
-  };
-  passwordRecoveryForm: {
-    loginId: string;
-    recoveryToken: string;
-    newPassword: string;
-  };
-  passwordRecoveryResult: PasswordRecoveryRequestResult | null;
-  passwordResetForm: {
-    currentPassword: string;
-    newPassword: string;
-  };
-  session: CustomerSession | null;
-  sessions: AuthSessionItem[];
-  totpChallengeForm: {
-    totpCode: string;
-    rememberDevice: boolean;
-  };
-  totpCode: string;
-  totpEnrollment: TotpEnrollmentStartResponse | null;
-  onBackupChallengeChange: (value: {
-    backupCode: string;
-    rememberDevice: boolean;
-  }) => void;
-  onBackupChallenge: (event: FormEvent<HTMLFormElement>) => void;
-  onDisableTotp: () => void;
-  onIssueBackupCodes: () => void;
-  onLoadSessions: () => void;
-  onLogin: (event: FormEvent<HTMLFormElement>) => void;
-  onLoginChange: (value: { loginId: string; password: string }) => void;
-  onLogout: () => void;
-  onPasswordRecoveryChange: (value: {
-    loginId: string;
-    recoveryToken: string;
-    newPassword: string;
-  }) => void;
-  onPasswordRecoveryConfirm: (event: FormEvent<HTMLFormElement>) => void;
-  onPasswordRecoveryRequest: (event: FormEvent<HTMLFormElement>) => void;
-  onPasswordReset: (event: FormEvent<HTMLFormElement>) => void;
-  onPasswordResetChange: (value: {
-    currentPassword: string;
-    newPassword: string;
-  }) => void;
-  onRefresh: () => void;
-  onRevokeAllSessions: () => void;
-  onRevokeSession: (sessionId: number) => void;
-  onStartTotpEnrollment: () => void;
-  onTotpChallenge: (event: FormEvent<HTMLFormElement>) => void;
-  onTotpChallengeChange: (value: { totpCode: string; rememberDevice: boolean }) => void;
-  onTotpCodeChange: (value: string) => void;
-  onVerifyTotpEnrollment: () => void;
-}) {
-  return (
-    <section className="task-section security-section">
-      <div className="section-title">
-        <div>
-          <p>인증센터</p>
-          <h1>로그인 및 보안관리</h1>
-        </div>
-        <div className="button-row">
-          <button disabled={!props.session || props.isBusy} onClick={props.onRefresh} type="button">
-            토큰 재발급
-          </button>
-          <button disabled={!props.session || props.isBusy} onClick={props.onLogout} type="button">
-            로그아웃
-          </button>
-        </div>
-      </div>
-
-      <div className="two-column">
-        <form className="bank-form" onSubmit={props.onLogin}>
-          <div className="form-heading">
-            <strong>로그인</strong>
-            <span>아이디와 비밀번호</span>
-          </div>
-          <label>
-            <span>이용자 ID</span>
-            <input
-              autoComplete="username"
-              onChange={(event) =>
-                props.onLoginChange({
-                  ...props.loginForm,
-                  loginId: event.target.value,
-                })
-              }
-              required
-              value={props.loginForm.loginId}
-            />
-          </label>
-          <label>
-            <span>비밀번호</span>
-            <input
-              autoComplete="current-password"
-              onChange={(event) =>
-                props.onLoginChange({
-                  ...props.loginForm,
-                  password: event.target.value,
-                })
-              }
-              required
-              type="password"
-              value={props.loginForm.password}
-            />
-          </label>
-          <button disabled={props.isBusy} type="submit">
-            로그인
-          </button>
-        </form>
-
-        <div className="bank-form passive">
-          <div className="form-heading">
-            <strong>현재 세션</strong>
-            <span>{props.session ? "로그인됨" : "로그인 필요"}</span>
-          </div>
-          <dl className="detail-list">
-            <div>
-              <dt>User ID</dt>
-              <dd>{props.session?.userId ?? "-"}</dd>
-            </div>
-            <div>
-              <dt>Access Token</dt>
-              <dd>{props.session ? maskToken(props.session.accessToken) : "-"}</dd>
-            </div>
-            <div>
-              <dt>Refresh 만료</dt>
-              <dd>{formatDateTime(props.session?.refreshExpiresAt)}</dd>
-            </div>
-          </dl>
-        </div>
-      </div>
-
-      {props.challenge ? (
-        <div className="two-column">
-          <form className="bank-form" onSubmit={props.onTotpChallenge}>
-            <div className="form-heading">
-              <strong>TOTP 추가 인증</strong>
-              <span>{props.challenge.challengeType ?? "TOTP"}</span>
-            </div>
-            <label>
-              <span>인증번호 6자리</span>
-              <input
-                inputMode="numeric"
-                maxLength={6}
-                onChange={(event) =>
-                  props.onTotpChallengeChange({
-                    ...props.totpChallengeForm,
-                    totpCode: event.target.value,
-                  })
-                }
-                required
-                value={props.totpChallengeForm.totpCode}
-              />
-            </label>
-            <label className="checkbox-line">
-              <input
-                checked={props.totpChallengeForm.rememberDevice}
-                onChange={(event) =>
-                  props.onTotpChallengeChange({
-                    ...props.totpChallengeForm,
-                    rememberDevice: event.target.checked,
-                  })
-                }
-                type="checkbox"
-              />
-              <span>이 기기 기억</span>
-            </label>
-            <button disabled={props.isBusy} type="submit">
-              인증
-            </button>
-          </form>
-
-          <form className="bank-form" onSubmit={props.onBackupChallenge}>
-            <div className="form-heading">
-              <strong>Backup Code 인증</strong>
-              <span>대체 인증수단</span>
-            </div>
-            <label>
-              <span>Backup Code</span>
-              <input
-                onChange={(event) =>
-                  props.onBackupChallengeChange({
-                    ...props.backupChallengeForm,
-                    backupCode: event.target.value,
-                  })
-                }
-                required
-                value={props.backupChallengeForm.backupCode}
-              />
-            </label>
-            <label className="checkbox-line">
-              <input
-                checked={props.backupChallengeForm.rememberDevice}
-                onChange={(event) =>
-                  props.onBackupChallengeChange({
-                    ...props.backupChallengeForm,
-                    rememberDevice: event.target.checked,
-                  })
-                }
-                type="checkbox"
-              />
-              <span>이 기기 기억</span>
-            </label>
-            <button disabled={props.isBusy} type="submit">
-              인증
-            </button>
-          </form>
-        </div>
-      ) : null}
-
-      <div className="two-column">
-        <form className="bank-form" onSubmit={props.onPasswordRecoveryRequest}>
-          <div className="form-heading">
-            <strong>비밀번호 찾기</strong>
-            <span>복구 요청</span>
-          </div>
-          <label>
-            <span>이용자 ID</span>
-            <input
-              onChange={(event) =>
-                props.onPasswordRecoveryChange({
-                  ...props.passwordRecoveryForm,
-                  loginId: event.target.value,
-                })
-              }
-              required
-              value={props.passwordRecoveryForm.loginId}
-            />
-          </label>
-          <button disabled={props.isBusy} type="submit">
-            복구 요청
-          </button>
-          {props.passwordRecoveryResult?.handoffRequestId ? (
-            <p className="field-note">
-              Request ID: {props.passwordRecoveryResult.handoffRequestId}
-            </p>
-          ) : null}
-        </form>
-
-        <form className="bank-form" onSubmit={props.onPasswordRecoveryConfirm}>
-          <div className="form-heading">
-            <strong>복구 확정</strong>
-            <span>토큰 기반 변경</span>
-          </div>
-          <label>
-            <span>복구 토큰</span>
-            <input
-              onChange={(event) =>
-                props.onPasswordRecoveryChange({
-                  ...props.passwordRecoveryForm,
-                  recoveryToken: event.target.value,
-                })
-              }
-              required
-              value={props.passwordRecoveryForm.recoveryToken}
-            />
-          </label>
-          <label>
-            <span>새 비밀번호</span>
-            <input
-              onChange={(event) =>
-                props.onPasswordRecoveryChange({
-                  ...props.passwordRecoveryForm,
-                  newPassword: event.target.value,
-                })
-              }
-              required
-              type="password"
-              value={props.passwordRecoveryForm.newPassword}
-            />
-          </label>
-          <button disabled={props.isBusy} type="submit">
-            변경
-          </button>
-        </form>
-      </div>
-
-      <div className="two-column">
-        <form className="bank-form" onSubmit={props.onPasswordReset}>
-          <div className="form-heading">
-            <strong>비밀번호 변경</strong>
-            <span>로그인 세션 필요</span>
-          </div>
-          <label>
-            <span>현재 비밀번호</span>
-            <input
-              onChange={(event) =>
-                props.onPasswordResetChange({
-                  ...props.passwordResetForm,
-                  currentPassword: event.target.value,
-                })
-              }
-              required
-              type="password"
-              value={props.passwordResetForm.currentPassword}
-            />
-          </label>
-          <label>
-            <span>새 비밀번호</span>
-            <input
-              onChange={(event) =>
-                props.onPasswordResetChange({
-                  ...props.passwordResetForm,
-                  newPassword: event.target.value,
-                })
-              }
-              required
-              type="password"
-              value={props.passwordResetForm.newPassword}
-            />
-          </label>
-          <button disabled={!props.session || props.isBusy} type="submit">
-            변경
-          </button>
-        </form>
-
-        <div className="bank-form passive">
-          <div className="form-heading">
-            <strong>MFA 관리</strong>
-            <span>TOTP / Backup Code</span>
-          </div>
-          <label>
-            <span>TOTP 인증번호</span>
-            <input
-              inputMode="numeric"
-              maxLength={6}
-              onChange={(event) => props.onTotpCodeChange(event.target.value)}
-              value={props.totpCode}
-            />
-          </label>
-          <div className="button-row compact">
-            <button
-              disabled={!props.session || props.isBusy}
-              onClick={props.onStartTotpEnrollment}
-              type="button"
-            >
-              등록 시작
-            </button>
-            <button
-              disabled={!props.session || props.isBusy}
-              onClick={props.onVerifyTotpEnrollment}
-              type="button"
-            >
-              등록 확인
-            </button>
-            <button
-              disabled={!props.session || props.isBusy}
-              onClick={props.onIssueBackupCodes}
-              type="button"
-            >
-              코드 발급
-            </button>
-            <button
-              disabled={!props.session || props.isBusy}
-              onClick={props.onDisableTotp}
-              type="button"
-            >
-              해지
-            </button>
-          </div>
-          {props.totpEnrollment ? (
-            <dl className="detail-list compact-list">
-              <div>
-                <dt>Secret</dt>
-                <dd>{props.totpEnrollment.secretKey}</dd>
-              </div>
-              <div>
-                <dt>만료</dt>
-                <dd>{formatDateTime(props.totpEnrollment.expiresAt)}</dd>
-              </div>
-            </dl>
-          ) : null}
-          {props.backupCodes ? (
-            <div className="code-list">
-              {props.backupCodes.backupCodes.map((code) => (
-                <code key={code}>{code}</code>
-              ))}
-            </div>
-          ) : null}
-        </div>
-      </div>
-
-      <section className="table-panel">
-        <div className="panel-toolbar">
-          <div>
-            <strong>접속 세션</strong>
-            <span>현재 사용자 refresh token session</span>
-          </div>
-          <div className="button-row compact">
-            <button
-              disabled={!props.session || props.isBusy}
-              onClick={props.onLoadSessions}
-              type="button"
-            >
-              조회
-            </button>
-            <button
-              disabled={!props.session || props.isBusy}
-              onClick={props.onRevokeAllSessions}
-              type="button"
-            >
-              전체 해지
-            </button>
-          </div>
-        </div>
-        <div className="bank-table-wrap">
-          <table className="bank-table">
-            <caption>인증 세션 목록</caption>
-            <thead>
-              <tr>
-                <th>Session ID</th>
-                <th>상태</th>
-                <th>기기</th>
-                <th>IP</th>
-                <th>최종 사용</th>
-                <th>관리</th>
-              </tr>
-            </thead>
-            <tbody>
-              {props.sessions.length === 0 ? (
-                <tr>
-                  <td colSpan={6}>조회된 세션이 없습니다.</td>
-                </tr>
-              ) : (
-                props.sessions.map((item) => (
-                  <tr key={item.sessionId}>
-                    <td>{item.sessionId}</td>
-                    <td>{item.currentSession ? "현재 세션" : item.sessionStatus}</td>
-                    <td>{item.deviceName ?? "-"}</td>
-                    <td>{item.ipAddress ?? "-"}</td>
-                    <td>{formatDateTime(item.lastUsedAt)}</td>
-                    <td>
-                      <button
-                        disabled={props.isBusy}
-                        onClick={() => props.onRevokeSession(item.sessionId)}
-                        type="button"
-                      >
-                        해지
-                      </button>
-                    </td>
-                  </tr>
-                ))
-              )}
-            </tbody>
-          </table>
-        </div>
-      </section>
-    </section>
   );
 }
