+++ front/src/components/customer-banking/common.tsx
@@ -1,0 +1,22 @@
+export function ResultPanel({ title, rows }: { title: string; rows: string[][] }) {
+  return (
+    <div className="result-panel">
+      <div className="form-heading">
+        <strong>{title}</strong>
+        <span>{rows.length > 0 ? "완료" : "대기"}</span>
+      </div>
+      {rows.length === 0 ? (
+        <p className="rail-copy">처리 결과가 여기에 표시됩니다.</p>
+      ) : (
+        <dl className="detail-list">
+          {rows.map(([key, value]) => (
+            <div key={key}>
+              <dt>{key}</dt>
+              <dd>{value}</dd>
+            </div>
+          ))}
+        </dl>
+      )}
+    </div>
+  );
+}
